requirements.txt
pygame-ce>=2.4.0
numpy>=1.26.0
watchdog>=4.0.0
pygame-gui>=0.6.9
